+++ Kernel/Output/HTML/LayoutTicket.pm
@@ -225,6 +225,7 @@ sub AgentQueueListOption {
 
     # set OnChange if AJAX is used
     if ( $Param{Ajax} ) {
+
         if ( !$Param{Ajax}->{Depend} ) {
             $Self->{LogObject}->Log(
                 Priority => 'error',
@@ -286,6 +287,9 @@ sub AgentQueueListOption {
         $Data{$_} .= '::';
     }
 
+    # get HTML utils object
+    my $HTMLUtilsObject = $Self->{HTMLUtilsObject};
+
     # build selection string
     for ( sort { $Data{$a} cmp $Data{$b} } keys %Data ) {
         my @Queue = split( /::/, $Param{Data}->{$_} );
@@ -333,7 +337,7 @@ sub AgentQueueListOption {
                         my $DSpace               = '&nbsp;&nbsp;' x $Index;
                         my $OptionTitleHTMLValue = '';
                         if ($OptionTitle) {
-                            my $HTMLValue = $Self->{HTMLUtilsObject}->ToHTML(
+                            my $HTMLValue = $HTMLUtilsObject->ToHTML(
                                 String => $Queue[$Index],
                             );
                             $OptionTitleHTMLValue = ' title="' . $HTMLValue . '"';
@@ -354,11 +358,14 @@ sub AgentQueueListOption {
             my $String               = $Space . $Queue[-1];
             my $OptionTitleHTMLValue = '';
             if ($OptionTitle) {
-                my $HTMLValue = $Self->{HTMLUtilsObject}->ToHTML(
+                my $HTMLValue = $HTMLUtilsObject->ToHTML(
                     String => $Queue[-1],
                 );
                 $OptionTitleHTMLValue = ' title="' . $HTMLValue . '"';
             }
+            my $HTMLValue = $HTMLUtilsObject->ToHTML(
+                String => $_,
+            );
             if (
                 $SelectedID eq $_
                 || $Selected eq $Param{Data}->{$_}
@@ -367,7 +374,7 @@ sub AgentQueueListOption {
             {
                 $Param{MoveQueuesStrg}
                     .= '<option selected="selected" value="'
-                    . $_ . '"'
+                    . $HTMLValue . '"'
                     . $OptionTitleHTMLValue . '>'
                     . $String
                     . "</option>\n";
@@ -383,7 +390,7 @@ sub AgentQueueListOption {
             else {
                 $Param{MoveQueuesStrg}
                     .= '<option value="'
-                    . $_ . '"'
+                    . $HTMLValue . '"'
                     . $OptionTitleHTMLValue . '>'
                     . $String
                     . "</option>\n";